Split replication manager routes in separate modules.
