+++ s3/replication/manager/src/s3replicationmanager/job_routes.py
@@ -23,82 +23,11 @@ import json
 from urllib.parse import urlparse, parse_qs
 from s3replicationcommon.job import JobJsonEncoder
 from .prepare_job import PrepareReplicationJob
-from .subscribers import Subscribers
 
 _logger = logging.getLogger('s3replicationmanager')
 
 # Route table declaration
 routes = web.RouteTableDef()
-
-
-@routes.post('/subscribers')  # noqa: E302
-async def add_subscriber(request):
-    """Handler to add subscriber."""
-    # Get subscriber details from payload
-    subscriber = await request.json()
-    _logger.debug('API: POST /subscribers\nContent : {}'.format(subscriber))
-
-    subscribers_list = request.app['subscribers']
-
-    # Check if subscriber is already present
-    subscriber_obj = subscribers_list.add_subscriber(subscriber)
-    _logger.debug('Subscriber added  : {}'.format(
-        subscriber_obj.get_dictionary()))
-
-    return web.json_response(subscriber_obj.get_dictionary(),
-                             status=201)
-
-
-@routes.get('/subscribers')  # noqa: E302
-async def list_subscribers(request):
-    """Handler to list subscribers."""
-    _logger.debug('API: GET /subscribers')
-    subscribers = request.app['subscribers']
-
-    _logger.debug('Number of subscribers {}'.format(subscribers.count()))
-    return web.json_response(subscribers, dumps=Subscribers.dumps, status=200)
-
-
-@routes.get('/subscribers/{subscriber_id}')  # noqa: E302
-async def get_subscriber(request):
-    """Handler to get job attributes."""
-    subscriber_id = request.match_info['subscriber_id']
-    _logger.debug('API: GET /subscribers/{}'.format(subscriber_id))
-
-    subscribers = request.app['subscribers']
-
-    subscriber = subscribers.get_subscriber(subscriber_id)
-
-    if subscriber is not None:
-        _logger.debug('Subscriber found with subscriber_id : {} '.
-                      format(subscriber_id))
-        _logger.debug('Subscriber details : {} '.format(
-            subscriber.get_dictionary()))
-        return web.json_response(subscriber.get_dictionary(), status=200)
-    else:
-        _logger.debug('Subscriber missing with subscriber_id : {} '.
-                      format(subscriber_id))
-        return web.json_response(
-            {'ErrorResponse': 'Subscriber Not Found!'}, status=404)
-
-
-@routes.delete('/subscribers/{subscriber_id}')  # noqa: E302
-async def remove_subscriber(request):
-    """Handler to remove subscriber."""
-    subscribers = request.app['subscribers']
-
-    subscriber_id = (request.match_info['subscriber_id'])
-    _logger.debug('API: DELETE /subscribers/{}'.format(subscriber_id))
-
-    # Check if subscriber is already present
-    if subscribers.is_subscriber_present(subscriber_id):
-        subscriber = subscribers.remove_subscriber(subscriber_id)
-        _logger.debug('Subscriber removed  : {}'.format(
-            subscriber.get_dictionary()))
-        return web.json_response({'subscriber_id': subscriber_id}, status=204)
-    else:
-        return web.json_response(
-            {'ErrorResponse': 'Subscriber Not Found!'}, status=404)
 
 
 @routes.post('/jobs')  # noqa: E302
@@ -165,20 +94,20 @@ async def list_jobs(request):
 
     _logger.debug('API: GET /jobs\n Query: {}'.format(query))
 
-    progressing_jobs_obj = request.app['jobs_in_progress']
-    all_jobs_obj = request.app['all_jobs']
+    inprogress_jobs_list = request.app['jobs_in_progress']
+    all_jobs_list = request.app['all_jobs']
 
     # Return in progress jobs
     if 'inprogress' in query:
         _logger.debug('InProgress query param: {}'.format(query['inprogress']))
         return web.json_response(
-            {'inprogress jobs': json.dumps(
-                list(progressing_jobs_obj.get_keys()))})
+            {'jobs-inprogress': json.dumps(
+                list(inprogress_jobs_list.get_keys()))})
 
     # Return total job counts
     elif 'count' in query:
         return web.json_response(
-            {'count': progressing_jobs_obj.count() + all_jobs_obj.count()})
+            {'count': inprogress_jobs_list.count() + all_jobs_list.count()})
 
     # Return requested jobs
     elif 'prefetch' in query:
@@ -190,22 +119,22 @@ async def list_jobs(request):
         if request.app['subscribers'].is_subscriber_present(sub_id):
 
             # Remove prefetch_count jobs from all_jobs and add to inprogress
-
-            add_inprogress = all_jobs_obj.remove_jobs(prefetch_count)
-            progressing_jobs_obj.add_jobs(add_inprogress)
+            add_inprogress = all_jobs_list.remove_jobs(prefetch_count)
+            inprogress_jobs_list.add_jobs(add_inprogress)
 
             _logger.debug('jobs in progress : {}'.format(
-                list(progressing_jobs_obj.get_keys())))
-            return web.json_response(
-                {'Response': json.dumps(
-                    list(progressing_jobs_obj.get_keys()))})
+                list(inprogress_jobs_list.get_keys())))
+            return web.json_response({
+                json.dumps(list(inprogress_jobs_list.get_keys()))
+            })
 
         # Subscriber is not in the list
         else:
             return web.json_response({'ErrorResponse': 'Invalid subscriber'})
     else:
-        return web.json_response(
-            {'Response': json.dumps(list(all_jobs_obj.get_keys()))})
+        return web.json_response({
+            json.dumps(list(all_jobs_list.get_keys()))
+        })
 
 
 @routes.delete('/jobs/{job_id}')  # noqa: E302
